Refactor IdentityController to use future.join() for token granting and rename GlobalExceptionHandler package for better organization
